pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-sbc-reliability-lab"
version = "1.0.0"
description = "A reproducible benchmarking and reliability framework for AI inference on Raspberry Pi 5"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Manu Nicholas Jacob", email = "manunicholasjacob@gmail.com"},
    {name = "Edge SBC Reliability Lab Contributors"}
]
keywords = [
    "raspberry-pi",
    "edge-ai",
    "benchmarking",
    "inference",
    "reliability",
    "thermal",
    "sbc",
    "onnx",
    "tflite",
    "pytorch"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Benchmark",
    "Topic :: System :: Hardware",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0",
    "psutil>=5.8.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
onnx = ["onnxruntime>=1.10.0"]
tflite = ["tflite-runtime>=2.10.0"]
torch = ["torch>=2.0.0"]
all = [
    "onnxruntime>=1.10.0",
    "tflite-runtime>=2.10.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]

[project.scripts]
sbc-bench = "edge_sbc_reliability_lab.cli:main"
sbc-health-check = "edge_sbc_reliability_lab.scripts.pre_run_health_check:main"
sbc-analyze = "edge_sbc_reliability_lab.analysis.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/edge-sbc-reliability-lab"
Documentation = "https://github.com/yourusername/edge-sbc-reliability-lab#readme"
Repository = "https://github.com/yourusername/edge-sbc-reliability-lab"
Issues = "https://github.com/yourusername/edge-sbc-reliability-lab/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
